backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
pandas==2.3.0
numpy==1.26.4
tqdm==4.67.1
vaderSentiment==3.3.2
nltk==3.9.1
google-generativeai==0.8.5
google-api-python-client==2.172.0
python-dotenv==1.1.0